lab2/requirements.txt
colorama==0.4.6
pycryptodome==3.21.0
